+++ tests/controllers/batch2/OrderControllerTest.php
@@ -5,7 +5,7 @@ use Redooor\Redminportal\App\Models\Product;
 
 class OrderControllerTest extends BaseControllerTest
 {
-    use TraitSorterControllerTest;
+    use TraitSorterControllerTest, TraitSearcherControllerTest;
     
     /**
      * Contructor
@@ -37,6 +37,10 @@ class OrderControllerTest extends BaseControllerTest
         
         // For testing sort
         $this->sortBy = 'created_at';
+        
+        // For testing search
+        $this->searchable_field = 'payment_status';
+        $this->search_text = 'Completed';
         
         parent::__construct($page, $viewhas, $input);
     }